==> 01-CrashRecords.ipynb <==
import pandas as pd
import geopandas as gpd
import matplotlib.pyplot as plt
import numpy as np
from pyproj import CRS
import pathlib
from pathlib import Path
from shapely import wkt
from tqdm import tqdm

import math
import codecs
import osm2geojson
from shapely import wkt

import gzip
from xml.etree.ElementTree import Element, SubElement, Comment, tostring
import xml.etree.ElementTree as ET

# set the working directory
BASE_DIR = Path.cwd()
# print(BASE_DIR)

# read the updated SWITRS_crash file and update the ref_latlong column, if empty
df = pd.read_csv(BASE_DIR.parent.joinpath("Exported_Files","Road_Crashes_SWITRS_TIMS_matched_output_v3.4.csv"))

# Create  multiple columns and assign them default values
new_cols = ["Total_Crash","COUNT_Fatal", "COUNT_Severe_Injury", "COUNT_Visible_Injury", "COUNT_Other_Injury", "COUNT_PDO"]
new_vals = [1,0,0,0,0,0]
df = df.reindex(columns=df.columns.tolist() + new_cols)   # add empty cols
df[new_cols] = new_vals  # multi-column assignment works for existing cols

# assign values depending upon the collision severity
df.loc[df["COLLISION_SEVERITY"]==1, "COUNT_Fatal"] = 1
df.loc[df["COLLISION_SEVERITY"]==2, "COUNT_Severe_Injury"] = 1
df.loc[df["COLLISION_SEVERITY"]==3, "COUNT_Visible_Injury"] = 1
df.loc[df["COLLISION_SEVERITY"]==4, "COUNT_Other_Injury"] = 1
df.loc[df["COLLISION_SEVERITY"]==0, "COUNT_PDO"] = 1

df.loc[df["COLLISION_SEVERITY"]==1, "COUNT_Fatal"] = 1
df.loc[df["COLLISION_SEVERITY"]==2, "COUNT_Severe_Injury"] = 1
df.loc[df["COLLISION_SEVERITY"]==3, "COUNT_Visible_Injury"] = 1
df.loc[df["COLLISION_SEVERITY"]==4, "COUNT_Other_Injury"] = 1
df.loc[df["COLLISION_SEVERITY"]==0, "COUNT_PDO"] = 1

df.to_csv(BASE_DIR.parent.joinpath("Exported_Files","Road_Crashes_SWITRS_TIMS_matched_output_modified.csv"))
