==> 05.2 - CensusTract sample.ipynb <==
import pandas as pd
import geopandas as gpd
import matplotlib.pyplot as plt
import numpy as np
from pyproj import CRS
import pathlib
from pathlib import Path
from shapely import wkt
from tqdm import tqdm

import math
import codecs
import osm2geojson
from shapely import wkt

import gzip
from xml.etree.ElementTree import Element, SubElement, Comment, tostring
import xml.etree.ElementTree as ET

import os

# set the working directory
BASE_DIR = Path.cwd()
# define the exported folder path
# Check if folder exists
folder_path = pathlib.Path(BASE_DIR.parent.joinpath("Exported_Files","census_tract","agg_network"))
folder_path.mkdir(parents=True, exist_ok=True)
# print(BASE_DIR)

road2010 = pd.read_csv(BASE_DIR.parent.joinpath("Exported_Files","census_tract","agg_network","615","RoadNetwork_2010.csv"))
road2010["ACCIDENT_YEAR"] = 2010
road2016 = pd.read_csv(BASE_DIR.parent.joinpath("Exported_Files","census_tract","agg_network","615","RoadNetwork_2016.csv"))
road2016["ACCIDENT_YEAR"] = 2016
dfrdntwrk = pd.concat([road2010,road2016], ignore_index=True)
dfrdntwrk = dfrdntwrk.groupby(["ACCIDENT_YEAR","tractce10"]).agg({"Tot_Vol":"sum",
                                                                  "Tot_CAP":"sum",
                                                                  "OOS":"sum",
                                                                  "PUDO":"sum",
                                                                  "CSPD_1":"mean",
                                                                  "DISTANCE":"sum",
                                                                  "TNC_Tot_Vol":"sum",
                                                                  "NON_TNC_Tot_Vol":"sum",
                                                                  "Tot_Vol_yr":"sum",
                                                                  "Tot_CAP_yr":"sum",
                                                                  "OOS_yr":"sum",
                                                                  "PUDO_yr":"sum",
                                                                  "TNC_Tot_Vol_yr":"sum",
                                                                  "NON_TNC_Tot_Vol_yr":"sum",
                                                                  })

crash2010 = pd.read_csv(BASE_DIR.parent.joinpath("Exported_Files","census_tract","agg_network","615","RoadCrash_2010.csv"))
crash2016 = pd.read_csv(BASE_DIR.parent.joinpath("Exported_Files","census_tract","agg_network","615","RoadCrash_2016.csv"))
dfcrash = pd.concat([crash2010,crash2016], ignore_index=True)
dfcrash = dfcrash.groupby(["ACCIDENT_YEAR","tractce10"])["Total_Crash"].sum()


df = pd.merge(dfrdntwrk,dfcrash,how="inner",right_on=["ACCIDENT_YEAR","tractce10"],left_on=["ACCIDENT_YEAR","tractce10"])
df.to_csv(BASE_DIR.parent.joinpath("Exported_Files","census_tract","agg_network","615","combined.csv"))

crash2010 = pd.read_csv(BASE_DIR.parent.joinpath("Exported_Files","census_tract","agg_network","615","RoadCrash_2010.csv"))
crash2016 = pd.read_csv(BASE_DIR.parent.joinpath("Exported_Files","census_tract","agg_network","615","RoadCrash_2016.csv"))
dfcrash = pd.concat([crash2010,crash2016], ignore_index=True)
dfcrash = dfcrash.groupby(["ACCIDENT_YEAR","tractce10"])["Total_Crash"].sum()

df = pd.merge(dfrdntwrk,dfcrash,how="inner",right_on=["ACCIDENT_YEAR","tractce10"],left_on=["ACCIDENT_YEAR","tractce10"])

df.to_csv(BASE_DIR.parent.joinpath("Exported_Files","census_tract","agg_network","615","combined.csv"))

crash = np.random.randint(0,326,size=(200,1))

PUDO = np.random.randint(0,30000,size=(200,1))

VMT_mil = np.random.uniform(0,300,size=(200,1))

df = pd.DataFrame()
df["crash"]=crash
df["PUDO"]=PUDO
df["VMT_mil"] = VMT_mil

n = 384
df = pd.DataFrame(dict(
    A=np.random.randint(0, 326, size=n),
    B=np.random.randint(0,30000, size=n),
    C=np.random.uniform(0,300,size=n)
    ))


df = pd.DataFrame((zip(crash,PUDO,VMT_mil)), columns=["Crash","PUDO","VMT_mil"])

df
